Download Data from the Integrated Data Tool REDcap Project

download_integrated_data connects to the Integrated Data Tool REDcap project via an API token and downloads the data currently available in the case interviews report.

download_integrated_data(
  date = Sys.Date(),
  api_token = Sys.getenv("redcap_IDT_token"),
  directory = paste0("V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/",
    "Integrated data tool Case Interviews/data/"),
  new_file = paste0("integrated_data_", date, ".csv"),
  convert = FALSE,
  force = FALSE
)

Arguments

date
The date with which to name new_file

api_token
The API access token for the Integrated Data Tool REDcap project. This should be stored in a .Renviron file; see using_Renviron for details.

directory
A string specifying the directory in which to save the data; this should usually be left alone

new_file
A string specifying the file name to save the data under; this should always end in .csv and should usually be left alone

convert
A logical indicating whether to open in Excel for conversion to .xlsx format (or another file format)

force
A logical indicating whether to overwrite an existing file with the same date; the default is FALSE and should usually be left alone

Value

Invisibly returns NULL

Details

download_integrated_data downloads data as above and saves to the file directory and file specified in directory and new_file (by default, it saves to directory V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Integrated data tool Case Interviews/ and file integrated_data_YYYY-MM-DD.csv.) Only change this if you're doing some custom analysis; by default, load_integrated_data will look here for data to import.

To use download_integrated_data you'll need an API access key for the associated REDcap project. Please contact jesse.smith@shelbycountytn.gov or umar.kabir@shelbycountytn.gov for more information.

Best practice is to save this API access token in a separate file; see using_Renviron for details.